Moving the DI stuff to its own namespace, to prepare for some changes coming soon.
